Update 8-Bit Carry Lookahead Adder
The CLA is a design that it attempts to reduce the critical path delay of the carry.  Through the use of Boolean functions called generate and propagate, provided below, a block of circuitry consisting of AND and XOR gates is designed to calculate the final carry out bit much faster.  The carry propagate determines if the carry input bit can propagate to the next stage and the carry generate determines if the carry output bit is generated in the current stage.